Adding special representation beans for my jersey endpoints. Updating Resources to be more inline with jersey examples.  This conflicts with some OSGi principles but that's what we get when we try to use tech that hasn't been able to play nice yet. It's actually working somewhat well now, which is pretty sweet.

+++ org.praxis.blog/org.praxis.blog.dao/src/main/java/org/praxis/blog/dao/AbstractDao.java
@@ -34,7 +34,7 @@ public abstract class AbstractDao<E extends Entity> implements Dao<E> {
   @Override
   public List<E> list() {
     final List<E> list = new ArrayList<E>();
-    for( int i = 0; i < 5; i++ ) {
+    for( long i = 0; i < 5; i++ ) {
       final E entity = createEntity();
       entity.setId(i);
       list.add(entity);
@@ -56,7 +56,7 @@ public abstract class AbstractDao<E extends Entity> implements Dao<E> {
     E entity;
     try {
       entity = entityClass.newInstance();
-    } catch (final Exception e) {
+    } catch( final Exception e ) {
       throw new RuntimeException(e);
     }
     return entity;
